// src/app/features/CauHoiThuongGap/type.ts
export interface ItemCauHoiThuongGapProps {
  cauHoi: string;
  cauTraLoi: string;
  createdAt: string;
  loaiCauHoi: string;
  updatedAt: string;
  _id: string;
  idTopic: string;
  topic: {
    createdAt: string;
    name: string;
    order: number;
    type: string;
    updatedAt: string;
    _id: string;
  };
}